requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
azure-core==1.35.0
botbuilder-core==4.17.0
botbuilder-integration-aiohttp==4.17.0
botbuilder-schema==4.17.0
botframework-connector==4.17.0
botframework-streaming==4.17.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
frozenlist==1.7.0
idna==3.10
isodate==0.7.2
jsonpickle==1.4.2
msal==1.33.0
msrest==0.7.1
multidict==6.6.4
oauthlib==3.3.1
propcache==0.3.2
pycparser==2.22
PyJWT==2.10.1
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
